codex-pdf-compilier/requirements.txt
espn_api>=0.33.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
black>=24.0.0
ruff>=0.5.0
pytest>=8.0.0
